Revert "Revert "Full codegen for `logdet` (#3576)" (#3601)"
This reverts commit 67e4efbc9f3dd63f63ead416d050e08043e1f0c8.
