+++ torch_xla/csrc/ops/ops_lower_fn.cpp
@@ -1,5 +1,6 @@
 #include <torch_xla/csrc/generated/LazyIr.h>
 
+#include "tensorflow/compiler/xla/client/lib/logdet.h"
 #include "tensorflow/compiler/xla/client/lib/math.h"
 #include "torch_xla/csrc/elementwise.h"
 #include "torch_xla/csrc/helpers.h"
@@ -52,6 +53,11 @@ torch_xla::XlaOpVector Cosh::Lower(LoweringContext* loctx) const {
   return ReturnOp(xla::Cosh(xla_input), loctx);
 }
 
+torch_xla::XlaOpVector Logdet::Lower(LoweringContext* loctx) const {
+  xla::XlaOp xla_input = loctx->GetOutputOp(operand(0));
+  return ReturnOp(xla::LogDet(xla_input), loctx);
+}
+
 torch_xla::XlaOpVector Inverse::Lower(LoweringContext* loctx) const {
   xla::XlaOp xla_input = loctx->GetOutputOp(operand(0));
   return ReturnOp(BuildInverse(xla_input), loctx);
@@ -83,6 +89,13 @@ torch_xla::XlaOpVector Sinh::Lower(LoweringContext* loctx) const {
   xla::XlaOp xla_input = loctx->GetOutputOp(operand(0));
   return ReturnOp(xla::Sinh(xla_input), loctx);
 }
+
+/* Blocked on https://github.com/pytorch/xla/issues/3596 */
+// torch_xla::XlaOpVector Slogdet::Lower(LoweringContext* loctx) const {
+//   xla::XlaOp xla_input = loctx->GetOutputOp(operand(0));
+//   xla::SignAndLogDet result = xla::SLogDet(xla_input);
+//   return ReturnOps({result.sign, result.logdet}, loctx);
+// }
 
 torch_xla::XlaOpVector Tan::Lower(LoweringContext* loctx) const {
   xla::XlaOp xla_input = loctx->GetOutputOp(operand(0));
