+++ torch_xla/csrc/ops/ops_xla_shape_fn.cpp
@@ -1,5 +1,6 @@
 #include "torch_xla/csrc/ops/ops_xla_shape_fn.h"
 
+#include "tensorflow/compiler/xla/client/lib/logdet.h"
 #include "torch_xla/csrc/helpers.h"
 
 namespace torch_xla {
@@ -44,6 +45,16 @@ xla::Shape InverseOutputShape(const torch::lazy::Value& input) {
   return GetXlaShape(input);
 }
 
+xla::Shape LogdetOutputShape(const torch::lazy::Value& input) {
+  const xla::Shape& input_shape = GetXlaShape(input);
+  XLA_CHECK_GE(input_shape.rank(), 2) << input_shape;
+  // The input tensor is ...,N,N
+  xla::Shape logdet_shape(input_shape);
+  logdet_shape.DeleteDimension(input_shape.rank() - 1);
+  logdet_shape.DeleteDimension(input_shape.rank() - 2);
+  return logdet_shape;
+}
+
 xla::Shape MaximumOutputShape(const torch::lazy::Value& input,
                               const torch::lazy::Value& other) {
   auto lower_for_shape_fn =
@@ -70,6 +81,16 @@ xla::Shape SinOutputShape(const torch::lazy::Value& input) {
 xla::Shape SinhOutputShape(const torch::lazy::Value& input) {
   return GetXlaShape(input);
 }
+
+/* Blocked on https://github.com/pytorch/xla/issues/3596 */
+// xla::Shape SlogdetOutputShape(const torch::lazy::Value& input) {
+//   auto lower_for_shape_fn =
+//       [](absl::Span<const xla::XlaOp> operands) -> xla::XlaOp {
+//     xla::SignAndLogDet result = xla::SLogDet(operands[0]);
+//     return xla::Tuple(operands[0].builder(), {result.sign, result.logdet});
+//   };
+//   return InferOutputShape({GetXlaShape(input)}, lower_for_shape_fn);
+// }
 
 xla::Shape TanOutputShape(const torch::lazy::Value& input) {
   return GetXlaShape(input);
