+++ torch_xla/csrc/ops/ops_xla_shape_fn.h
@@ -23,6 +23,8 @@ xla::Shape CoshOutputShape(const torch::lazy::Value& input);
 
 xla::Shape InverseOutputShape(const torch::lazy::Value& input);
 
+xla::Shape LogdetOutputShape(const torch::lazy::Value& input);
+
 xla::Shape MaximumOutputShape(const torch::lazy::Value& input,
                               const torch::lazy::Value& other);
 
@@ -33,6 +35,9 @@ xla::Shape SignOutputShape(const torch::lazy::Value& input);
 xla::Shape SinOutputShape(const torch::lazy::Value& input);
 
 xla::Shape SinhOutputShape(const torch::lazy::Value& input);
+
+/* Blocked on https://github.com/pytorch/xla/issues/3596 */
+// xla::Shape SlogdetOutputShape(const torch::lazy::Value& input);
 
 xla::Shape TanOutputShape(const torch::lazy::Value& input);
